+++ src/client/index.js
@@ -6,7 +6,7 @@ import { match, browserHistory } from 'react-router';
 import { Root, Routes } from '../shared/components';
 import configureStore from '../shared/configureStore';
 import Immutable from 'immutable';
-import { decode } from '../shared/encoding';
+import { decode } from '../shared/lib/encoding';
 
 
 // NB: e30= is encoded '{}'.
@@ -17,9 +17,6 @@ export const load = (encoded = 'e30=') => {
     const store = configureStore(initialState);
 
     match({ routes: Routes, history: browserHistory }, (error, redirectLocation, renderProps) => {
-        // XXX This is certainly wrong, but unclear what the right approach is.
-        // Without forcing this location the page won't load.
-        // store.getState().routing.location = initialState.routing.location;
         render(
             <Root store={ store } {...renderProps} />,
             document.getElementById('root')
