+++ src/shared/components/Routes.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Route, IndexRoute } from 'react-router';
-import App from './App';
-import Home from './Home';
-import Post from './Post';
+import App from './views/App';
+import Home from './views/Home';
+import Post from './views/Post';
 
 export default (
     <Route path="/" component={ App }>
