+++ src/shared/components/index.js
@@ -1,6 +1,2 @@
 export { default as Root } from './Root';
-export { default as App } from './App';
-export { default as Home } from './Home';
-export { default as Post } from './Post';
-export { default as Page } from './Page';
 export { default as Routes } from './Routes';
