+++ src/shared/components/views/Home.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router';
-import { fetchIndexIfNecessary } from '../actions';
+import { fetchIndexIfNecessary } from '../../actions';
 import Immutable from 'immutable';
