+++ src/shared/components/views/Post.jsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { fetchPostIfNecessary } from '../actions';
+import { fetchPostIfNecessary } from '../../actions';
 
 class Post extends Component {
