+++ src/shared/lib/encoding/decode.js
