+++ src/shared/lib/encoding/encode.js
