+++ src/shared/lib/encoding/index.js
